new version

+++ assets/prepare-data.sh
@@ -2,19 +2,20 @@
 
 ###################################
 # Descobrir caractere hexa no vim: ga
-# substituir caractere hexa no vim: :%s/\%x<hex-value>//g
+# Substituir caractere hexa no vim: :%s/\%x<hex-value>//g
 #######################################
 
 BASE_PATH="`pwd`" # Preencha esse valor com um caminho para ler os arquivos xls e converter para csv.
 
 declare -a crimes=(
+	'Feminicidio'
 	'FurtoCelular'
 	'FurtoVeiculo'
 	'Homicicio'
 	'Latrocinio'
 	'LesaoMorte'
 	'MortePolicial'
-	#'MorteSuspeita'
+	'MorteSuspeita'
 	'RouboCelular'
 	'RouboVeiculo'
 );
@@ -35,6 +36,8 @@ declare -a years=(
 	'2015'
 	'2016'
 	'2017'
+	'2018'
+	'2019'
 );
 
 for crime in ${crimes[@]}; do
@@ -48,39 +51,25 @@ for crime in ${crimes[@]}; do
 
 			lcrime=$(echo $crime | tr '[:upper:]' '[:lower:]')
 
-			cp -v $crime\_$year\_$month.xls ../../wrangled/$year/$lcrime\_$year\_$month.tsv
+			if [ -f $crime\_$year\_$month.xls ]; then
+				cp -v $crime\_$year\_$month.xls ../../wrangled/$year/$lcrime\_$year\_$month.tsv
 
-			pushd ../../wrangled/$year > /dev/null
-			sed -i 's/\x00//g' $lcrime\_$year\_$month.tsv
-			sed -i 's/,//g' $lcrime\_$year\_$month.tsv
-			sed -i 's/\xc7/C/g' $lcrime\_$year\_$month.tsv
-			sed -i 's/\xfa/u/g' $lcrime\_$year\_$month.tsv
-			sed -i 's/\xed/i/g' $lcrime\_$year\_$month.tsv
-			sed -i 's/\xf7/o/g' $lcrime\_$year\_$month.tsv
-			sed -i 's/\xf4/o/g' $lcrime\_$year\_$month.tsv
-			sed -i 's/\xe3/a/g' $lcrime\_$year\_$month.tsv
-			sed -i 's/\xc3/A/g' $lcrime\_$year\_$month.tsv
-			sed -i 's/\xc1/A/g' $lcrime\_$year\_$month.tsv
-			sed -i 's/\xc2/A/g' $lcrime\_$year\_$month.tsv
-			sed -i 's/\xcd/I/g' $lcrime\_$year\_$month.tsv
-			sed -i 's/\xe0/a/g' $lcrime\_$year\_$month.tsv
-			sed -i 's/\xe9/e/g' $lcrime\_$year\_$month.tsv
-			sed -i 's/\xc9/E/g' $lcrime\_$year\_$month.tsv
-			sed -i 's/\xca/E/g' $lcrime\_$year\_$month.tsv
-			sed -i 's/\xea/e/g' $lcrime\_$year\_$month.tsv
-			sed -i 's/\xc9/E/g' $lcrime\_$year\_$month.tsv
-			sed -i 's/\xd3/O/g' $lcrime\_$year\_$month.tsv
-			sed -i 's/\xd4/O/g' $lcrime\_$year\_$month.tsv
-			sed -i 's/\xd5/O/g' $lcrime\_$year\_$month.tsv
-			sed -i 's/\xda/U/g' $lcrime\_$year\_$month.tsv
-			sed -i 's/\xaa/a./g' $lcrime\_$year\_$month.tsv
-			sed -i 's/\xe7/c/g' $lcrime\_$year\_$month.tsv
-			sed -i 's/\xa7/par. /g' $lcrime\_$year\_$month.tsv
-			sed -i 's/\xba/o./g' $lcrime\_$year\_$month.tsv
-			sed -i 's/\"//g' $lcrime\_$year\_$month.tsv
-			cat $lcrime\_$year\_$month.tsv | tr '\t' ',' >  $lcrime\_$year\_$month.csv
-			rm -rf $lcrime\_$year\_$month.tsv
-			popd > /dev/null
+				pushd ../../wrangled/$year > /dev/null
+				sed -i 's/\x00//g' $lcrime\_$year\_$month.tsv
+				sed -i 's/\"//g' $lcrime\_$year\_$month.tsv
+				cat $lcrime\_$year\_$month.tsv | tr '\t' ',' >  $lcrime\_$year\_$month.csv
+
+				# Descobre o encoding atual do arquivo para assim converter para UTF-8
+				ENCODING=$(file -i $lcrime\_$year\_$month.csv | cut -d ';' -f2 | cut -d '=' -f2 | tr '[:lower:]' '[:upper:]')
+				iconv -f $ENCODING -t UTF-8 $lcrime\_$year\_$month.csv -o $lcrime\_$year\_$month.csv.converted
+
+				# Efetua o processo de conversao para UTF-8, remove o arquivo antigo para ficar com o novo
+				rm -rf $lcrime\_$year\_$month.csv
+				mv $lcrime\_$year\_$month.csv.converted $lcrime\_$year\_$month.csv
+
+				rm -rf $lcrime\_$year\_$month.tsv
+				popd > /dev/null
+			fi
 		done
 		popd > /dev/null
 	done
